+++ NEWS.md
@@ -1,7 +1,9 @@
-# areal 0.1.3.9000
+# areal 0.1.3
 
 * Fix [Issue 6](https://github.com/slu-openGIS/areal/issues/6) - edge case where `st_intersection` creates a geometry collection
 * Fix [Issue 7](https://github.com/slu-openGIS/areal/issues/7) - interpolation returns `NA` values when `tid` and `sid` are the same variable name (e.g. both are `GEOID`)
+* Fix [Issue 14](https://github.com/slu-openGIS/areal/issues/14) - interpolation fails when the `sf` geometry column is not named `geometry` - now renamed on the fly if this is not the case
+* Fix [Issue 16](https://github.com/slu-openGIS/areal/issues/16) - tibble output missing other output variables
 
 * Full draft of JOSS paper manuscript added in `paper/`